add tests for dataformat
