+++ backend/backendTests/DataFormatControllerTests.cs
@@ -1,0 +1,89 @@
+﻿using OpenData.Domain.Models;
+using OpenData.Resources;
+using System.Net.Http;
+using System.Net.Http.Headers;
+using System.Text;
+using System.Text.Json;
+using System.Threading.Tasks;
+using Xunit;
+
+namespace OpenData.backend
+{
+    public class DataFormatControllerTests
+    : IClassFixture<CustomWebApplicationFactory<Startup>>
+    {
+
+        private readonly CustomWebApplicationFactory<Startup> _factory;
+
+        public DataFormatControllerTests(CustomWebApplicationFactory<Startup> factory)
+        {
+            _factory = factory;
+        }
+
+        /// <summary>
+        /// Asserts correct status code and correct content type for all HttpPut calls on the dataformat controller
+        /// </summary>
+        [Theory]
+        [InlineData("/api/dataformat")]
+        public async Task Put_NewDataFormat_EndpointsReturnSuccess_CorrectContentType(string url)
+        {
+            // Arrange
+
+            var client = _factory.CreateDefaultClient();
+            NewUserResource newUserResource = new NewUserResource();
+            newUserResource.Mail = "PutMet@test.kommune.no";
+            newUserResource.Password = "PutMet@passw0rd";
+            newUserResource.UserType = UserType.Municipality;
+            await client.PutAsync("api/user", new StringContent(JsonSerializer.Serialize(newUserResource), Encoding.UTF8, "application/json"));
+
+            AuthUserResource loginResource = new AuthUserResource();
+            loginResource.Mail = newUserResource.Mail;
+            loginResource.Password = newUserResource.Password;
+            var loginResponse = await client.PostAsync("/api/user/auth", new StringContent(JsonSerializer.Serialize(loginResource), Encoding.UTF8, "application/json"));
+            Assert.NotNull(loginResponse.Content);
+            var user = ResponseSerializer.Extract<PrivateSafeUserResource>(loginResponse);
+            Assert.Equal(loginResource.Mail, user.Mail);
+            string token = user.Token;
+            client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+
+            // Act
+
+            NewDataFormatResource resource = new NewDataFormatResource();
+            resource.Name = "Data Resource";
+            resource.MimeType = "Mime Type";
+            resource.Description = "Some DEscription...";
+            resource.DocumentationUrl = "URL";
+            var response = await client.PutAsync(url, new StringContent(JsonSerializer.Serialize(resource), Encoding.UTF8, "application/json"));
+
+            // Assert
+
+            response.EnsureSuccessStatusCode(); // Status Code 200-299
+            Assert.Equal("application/json; charset=utf-8",
+                response.Content.Headers.ContentType.ToString());
+
+        }
+
+        /// <summary>
+        /// Asserts correct status code and correct content type for all HttpGet calls on the dataformat controller
+        /// </summary>
+        [Theory]
+        [InlineData("/api/dataformat")]
+        public async Task Get_EndpointsReturnSuccess_CorrectContentType(string url)
+        {
+            // Arrange
+
+            var client = _factory.CreateDefaultClient();
+
+            // Act
+
+            var response = await client.GetAsync(url);
+
+            // Assert
+
+            response.EnsureSuccessStatusCode(); // Status Code 200-299
+            Assert.Equal("application/json; charset=utf-8",
+                response.Content.Headers.ContentType.ToString());
+        }
+
+    }
+}
