+++ backend/backendTests/MetadataControllerTests.cs
@@ -21,45 +21,68 @@ namespace OpenData.backend
         }
 
         /// <summary>
-        /// Asserts correct status code and correct content type for all HttpPut calls on the metadata controller
+        /// Asserts correct status code and correct content type for HttpPut calls on the metadata controller (for new metadata)
         /// </summary>
-        [Theory]
-        [InlineData("/api/metadata")]
-        public async Task Put_NewReleasedMetaData_Get_EndpointsReturnSuccess_CorrectContentType(string url)
-        {
-            // Arrange
+        //[Theory]
+        //[InlineData("/api/metadata")]
+        //public async Task Put_NewUnreleasedMetaData_EndpointsReturnSuccess_CorrectContentType(string url)
+        //{
+        //    // Arrange
 
-            var client = _factory.CreateDefaultClient();
-            NewUserResource newUserResource = new NewUserResource();
-            newUserResource.Mail = "PutMet@test.kommune.no";
-            newUserResource.Password = "PutMet@passw0rd";
-            newUserResource.UserType = UserType.Municipality;
-            await client.PutAsync("api/user", new StringContent(JsonSerializer.Serialize(newUserResource), Encoding.UTF8, "application/json"));
+        //    var client = _factory.CreateDefaultClient();
+        //    NewUserResource newUserResource = new NewUserResource();
+        //    newUserResource.Mail = "PutMet@test.kommune.no";
+        //    newUserResource.Password = "PutMet@passw0rd";
+        //    newUserResource.UserType = UserType.Municipality;
+        //    await client.PutAsync("api/user", new StringContent(JsonSerializer.Serialize(newUserResource), Encoding.UTF8, "application/json"));
 
-            AuthUserResource loginResource = new AuthUserResource();
-            loginResource.Mail = newUserResource.Mail;
-            loginResource.Password = newUserResource.Password;
-            var loginResponse = await client.PostAsync("/api/user/auth", new StringContent(JsonSerializer.Serialize(loginResource), Encoding.UTF8, "application/json"));
-            Assert.NotNull(loginResponse.Content);
-            var user = ResponseSerializer.Extract<PrivateSafeUserResource>(loginResponse);
-            Assert.Equal(loginResource.Mail, user.Mail);
-            string token = user.Token;
-            client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
-            SaveMetadataResource resource = new SaveMetadataResource();
-            resource.Description = "Test Description";
-            resource.MunicipalityName = "Test";
-            resource.ReleaseState = EReleaseState.Released;
+        //    AuthUserResource loginResource = new AuthUserResource();
+        //    loginResource.Mail = newUserResource.Mail;
+        //    loginResource.Password = newUserResource.Password;
+        //    var loginResponse = await client.PostAsync("/api/user/auth", new StringContent(JsonSerializer.Serialize(loginResource), Encoding.UTF8, "application/json"));
+        //    Assert.NotNull(loginResponse.Content);
+        //    var user = ResponseSerializer.Extract<PrivateSafeUserResource>(loginResponse);
+        //    Assert.Equal(loginResource.Mail, user.Mail);
+        //    string token = user.Token;
+        //    client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
 
-            // Act
+        //    SaveMetadataResource metaDataResource = new SaveMetadataResource();
+        //    metaDataResource.Description = "Test Description";
+        //    metaDataResource.MunicipalityName = "Test";
+        //    metaDataResource.ReleaseState = EReleaseState.Released;
 
-            var response = await client.PutAsync(url, new StringContent(JsonSerializer.Serialize(resource), Encoding.UTF8, "application/json"));
+        //    // Act
 
-            // Assert
+        //    var metaDataResponse = await client.PutAsync(url, new StringContent(JsonSerializer.Serialize(metaDataResource), Encoding.UTF8, "application/json"));
+        //    var metadata = ResponseSerializer.Extract<Metadata>(metaDataResponse);
+        //    Assert.Equal(user.MunicipalityName, metadata.MunicipalityName);
+        //    NewDataFormatResource dataFormatResource = new NewDataFormatResource();
+        //    var dataFormatResponse = await client.PutAsync("/api/dataformat", new StringContent(JsonSerializer.Serialize(dataFormatResource), Encoding.UTF8, "application/json"));
+        //    var dataformat = ResponseSerializer.Extract<DataFormat>(dataFormatResponse);
 
-            response.EnsureSuccessStatusCode(); // Status Code 200-299
-            Assert.Equal("application/json; charset=utf-8",
-                response.Content.Headers.ContentType.ToString());
-        }
+        //    NewDataSourceResource dataSourceResource = new NewDataSourceResource();
+        //    dataSourceResource.MetadataUuid = metadata.Uuid;
+        //    dataSourceResource.Url = "";
+        //    dataSourceResource.Description = "Some test source";
+        //    dataSourceResource.StartDate = System.DateTime.MinValue;
+        //    dataSourceResource.EndDate = System.DateTime.MaxValue;
+        //    dataSourceResource.DataFormatMimeType = dataformat.MimeType;
+        //    Assert.NotEqual(dataSourceResource.StartDate, dataSourceResource.EndDate);
+
+        //    var dataSourceResponse = await client.PutAsync(url + "/url", new StringContent(JsonSerializer.Serialize(dataSourceResource), Encoding.UTF8, "application/json"));
+
+        //    // Assert
+
+        //    metaDataResponse.EnsureSuccessStatusCode(); // Status Code 200-299
+        //    Assert.Equal("application/json; charset=utf-8",
+        //        metaDataResponse.Content.Headers.ContentType.ToString());
+
+        //    dataSourceResponse.EnsureSuccessStatusCode(); // Status Code 200-299
+        //    Assert.Equal("application/json; charset=utf-8",
+        //        dataSourceResponse.Content.Headers.ContentType.ToString());
+
+        //}
+
 
         /// <summary>
         /// Asserts correct status code and correct content type for all HttpGet calls on the metadata controller
